requirements.txt
fastapi
hypercorn
httpx
openai
langchain
langchain-openai
langchain-community
python-dotenv